Modify Pipe to return an RRef.
As per proposal in https://github.com/pytorch/pytorch/issues/44827,
the API needs to return an RRef to support inter-host pipelining.

For now, we just return a local RRef and only support pipeline on a single
host. But having this change in the API upfront ensures we don't make any BC
breaking changes later.

Differential Revision: [D24914022](https://our.internmc.facebook.com/intern/diff/D24914022/)

ghstack-source-id: 116499310
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47829